fix: ignore updates that happen before the full state is send
This can happen when the kernel is updating a widget from a thread
at the same time as when we are requesting the states.

When this happens, we get an update without the full widget state, which
causes the widget creation fo fail (since model and module are all undefined)

Note that this does not affect the new faster way, so ipywidgets 8 should
not be affected by this bug.